requirements.txt
flask==2.0.1
Pillow==9.0.0
python-dotenv==0.19.0
requests==2.26.0
pandas==1.3.3
numpy==1.21.2
opencv-python==4.5.3.56
torch==1.13.1
torchvision==0.14.1
geopy>=2.3.0
xlsxwriter>=3.0.0 